refactor: move theme files to designsystem module
