+++ core/presentation/designsystem/src/desktopMain/kotlin/com/parksupark/soomjae/core/presentation/designsystem/theme/Theme.jvm.kt
@@ -1,4 +1,4 @@
-package com.parksupark.soomjae.theme
+package com.parksupark.soomjae.core.presentation.designsystem.theme
 
 import androidx.compose.runtime.Composable
