+++ core/presentation/designsystem/src/iosMain/kotlin/com/parksupark/soomjae/core/presentation/designsystem/theme/Theme.ios.kt
@@ -1,4 +1,4 @@
-package com.parksupark.soomjae.theme
+package com.parksupark.soomjae.core.presentation.designsystem.theme
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
